Add http context accessor implementation

+++ Services/CharacterService/CharacterService.cs
@@ -39,14 +39,21 @@ namespace MYAPP.Services.CharacterService
 
             try
             {
+                //Adding character via data context
                 Character character = _mapper.Map<Character>(newCharacter);
-                _context.Characters.Add(character);
-                await _context.SaveChangesAsync();
+                character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
-                var dbCharacter = await _context.Characters.ToListAsync();
-                serviceResponse.Data = dbCharacter
+                _context.Characters.Add(character);
+                await _context.SaveChangesAsync();//saving changes to database
+
+                //display characters that belongs to the current user 
+                var dbCharacter = await _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
                     .Select(c => _mapper.Map<GetCharacterDtoResponse>(c))
+                    .ToListAsync();
+                var sortedCharacterList = dbCharacter
                     .OrderBy(c => c.Id).ToList();
+                serviceResponse.Data = sortedCharacterList;
             }
             catch (Exception exc)
             {
@@ -62,16 +69,29 @@ namespace MYAPP.Services.CharacterService
             ServiceResponse<List<GetCharacterDtoResponse>> response = new ServiceResponse<List<GetCharacterDtoResponse>>();
             
             try{// If the character exist
-                Character character = await _context.Characters.FirstAsync(c => c.Id == deleteCharacterRequest.Id); 
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                
-                var dbCharacter = await _context.Characters.ToListAsync();
-                response.Data = dbCharacter
-                    .Select(c => _mapper.Map<GetCharacterDtoResponse>(c))
-                    .OrderBy(c => c.Id).ToList();
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == deleteCharacterRequest.Id && c.User.Id == GetUserId()); 
 
-                response.Message = "Character ID:" + deleteCharacterRequest.Id + " has been deleted";
+                if(character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    
+                    var dbCharacter = await _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDtoResponse>(c))
+                        .ToListAsync();
+                    response.Data = dbCharacter
+                        .OrderBy(c => c.Id).ToList();
+                        
+                    response.Message = "Character ID:" + deleteCharacterRequest.Id + " has been deleted";
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found!";
+                }
+               
             }
             catch (Exception exc)
             {
@@ -95,10 +115,11 @@ namespace MYAPP.Services.CharacterService
             {
                 var dbCharacter = await _context.Characters
                     .Where(c => c.User.Id == GetUserId())
-                    .ToListAsync();
-                response.Data = dbCharacter
                     .Select(c => _mapper.Map<GetCharacterDtoResponse>(c))
+                    .ToListAsync();
+                var sortedCharacterList = dbCharacter
                     .OrderBy(c => c.Id).ToList();
+                response.Data = sortedCharacterList;
             }
             catch (Exception exc)
             {
@@ -116,7 +137,8 @@ namespace MYAPP.Services.CharacterService
 
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == singleCharacterRequest.Id);
+                var dbCharacter = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == singleCharacterRequest.Id && c.User.Id == GetUserId());
                 serviceResponse.Data = _mapper.Map<GetCharacterDtoResponse>(dbCharacter);
     
                 //If Characater not found
